# internal/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a scraper API server.",
        "title": "wormface server",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/twitter/{userName}/posts": {
            "get": {
                "description": "Get a list of tweets for a given user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "twitter"
                ],
                "summary": "Get user tweets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Twitter username",
                        "name": "userName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/twitter.TweetEntry"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/youtube/{channelName}/videos": {
            "get": {
                "description": "Get a list of videos for a given YouTube channel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "youtube"
                ],
                "summary": "Get channel videos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YouTube channel name",
                        "name": "channelName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/youtube.YouTubeVideo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "twitter.Description": {
            "type": "object",
            "properties": {
                "urls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/twitter.URLElement"
                    }
                }
            }
        },
        "twitter.Hashtag": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
        "twitter.Media": {
            "type": "object",
            "properties": {
                "display_url": {
                    "type": "string"
                },
                "expanded_url": {
                    "type": "string"
                },
                "id_str": {
                    "type": "string"
                },
                "media_url_https": {
                    "type": "string"
                },
                "original_info": {
                    "$ref": "#/definitions/twitter.OriginalInfo"
                },
                "sizes": {
                    "$ref": "#/definitions/twitter.Sizes"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "video_info": {
                    "type": "object",
                    "properties": {
                        "aspect_ratio": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        },
                        "duration_millis": {
                            "type": "integer"
                        },
                        "variants": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "content_type": {
                                        "type": "string"
                                    },
                                    "url": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "twitter.OriginalInfo": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "integer"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "twitter.PurpleEntities": {
            "type": "object",
            "properties": {
                "description": {
                    "$ref": "#/definitions/twitter.Description"
                },
                "url": {
                    "$ref": "#/definitions/twitter.Description"
                }
            }
        },
        "twitter.Sizes": {
            "type": "object",
            "properties": {
                "large": {
                    "$ref": "#/definitions/twitter.ThumbClass"
                },
                "medium": {
                    "$ref": "#/definitions/twitter.ThumbClass"
                },
                "small": {
                    "$ref": "#/definitions/twitter.ThumbClass"
                },
                "thumb": {
                    "$ref": "#/definitions/twitter.ThumbClass"
                }
            }
        },
        "twitter.ThumbClass": {
            "type": "object",
            "properties": {
                "h": {
                    "type": "integer"
                },
                "resize": {
                    "type": "string"
                },
                "w": {
                    "type": "integer"
                }
            }
        },
        "twitter.TimelineTweetItem": {
            "type": "object",
            "properties": {
                "itemType": {
                    "type": "string"
                },
                "tweet_results": {
                    "type": "object",
                    "properties": {
                        "result": {
                            "$ref": "#/definitions/twitter.TimelineTweetResult"
                        }
                    }
                }
            }
        },
        "twitter.TimelineTweetLegacy": {
            "type": "object",
            "properties": {
                "bookmark_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "entities": {
                    "$ref": "#/definitions/twitter.TweetEntities"
                },
                "favorite_count": {
                    "type": "integer"
                },
                "full_text": {
                    "type": "string"
                },
                "id_str": {
                    "type": "string"
                },
                "is_quote_status": {
                    "type": "boolean"
                },
                "quote_count": {
                    "type": "integer"
                },
                "reply_count": {
                    "type": "integer"
                },
                "retweet_count": {
                    "type": "integer"
                },
                "retweeted_status_result": {
                    "type": "object",
                    "properties": {
                        "result": {
                            "$ref": "#/definitions/twitter.TimelineTweetResult"
                        }
                    }
                },
                "user_id_str": {
                    "type": "string"
                }
            }
        },
        "twitter.TimelineTweetResult": {
            "type": "object",
            "properties": {
                "core": {
                    "type": "object",
                    "properties": {
                        "user_results": {
                            "$ref": "#/definitions/twitter.UserResult"
                        }
                    }
                },
                "legacy": {
                    "$ref": "#/definitions/twitter.TimelineTweetLegacy"
                },
                "rest_id": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "views": {
                    "$ref": "#/definitions/twitter.Views"
                }
            }
        },
        "twitter.TweetEntities": {
            "type": "object",
            "properties": {
                "hashtags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/twitter.Hashtag"
                    }
                },
                "media": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/twitter.Media"
                    }
                },
                "symbols": {
                    "type": "array",
                    "items": {}
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/twitter.URLElement"
                    }
                },
                "user_mentions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/twitter.UserMention"
                    }
                }
            }
        },
        "twitter.TweetEntry": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "object",
                    "properties": {
                        "entryType": {
                            "type": "string"
                        },
                        "itemContent": {
                            "$ref": "#/definitions/twitter.TimelineTweetItem"
                        },
                        "items": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "entryId": {
                                        "type": "string"
                                    },
                                    "item": {
                                        "type": "object",
                                        "properties": {
                                            "itemContent": {
                                                "$ref": "#/definitions/twitter.TimelineTweetItem"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "entryId": {
                    "type": "string"
                },
                "sortIndex": {
                    "type": "string"
                }
            }
        },
        "twitter.URLElement": {
            "type": "object",
            "properties": {
                "display_url": {
                    "type": "string"
                },
                "expanded_url": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "twitter.UserLegacy": {
            "type": "object",
            "properties": {
                "can_dm": {
                    "type": "boolean"
                },
                "can_media_tag": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "default_profile": {
                    "type": "boolean"
                },
                "default_profile_image": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "entities": {
                    "$ref": "#/definitions/twitter.PurpleEntities"
                },
                "fast_followers_count": {
                    "type": "integer"
                },
                "favourites_count": {
                    "type": "integer"
                },
                "followers_count": {
                    "type": "integer"
                },
                "following": {
                    "type": "boolean"
                },
                "friends_count": {
                    "type": "integer"
                },
                "has_custom_timelines": {
                    "type": "boolean"
                },
                "is_translator": {
                    "type": "boolean"
                },
                "listed_count": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "media_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "normal_followers_count": {
                    "type": "integer"
                },
                "pinned_tweet_ids_str": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "possibly_sensitive": {
                    "type": "boolean"
                },
                "profile_banner_url": {
                    "type": "string"
                },
                "profile_image_url_https": {
                    "type": "string"
                },
                "profile_interstitial_type": {
                    "type": "string"
                },
                "screen_name": {
                    "type": "string"
                },
                "statuses_count": {
                    "type": "integer"
                },
                "translator_type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                },
                "verified_type": {
                    "type": "string"
                },
                "want_retweets": {
                    "type": "boolean"
                },
                "withheld_in_countries": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "twitter.UserMention": {
            "type": "object",
            "properties": {
                "id_str": {
                    "type": "string"
                },
                "indices": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                },
                "screen_name": {
                    "type": "string"
                }
            }
        },
        "twitter.UserResult": {
            "type": "object",
            "properties": {
                "has_graduated_access": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "is_blue_verified": {
                    "type": "boolean"
                },
                "legacy": {
                    "$ref": "#/definitions/twitter.UserLegacy"
                },
                "parody_commentary_fan_label": {
                    "type": "string"
                },
                "profile_image_shape": {
                    "type": "string"
                },
                "rest_id": {
                    "type": "string"
                }
            }
        },
        "twitter.Views": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "youtube.YouTubeVideo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "publishedAt": {
                    "type": "string"
                },
                "thumbnailUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "videoId": {
                    "type": "string"
                }
            }
        }
    }
}